requirements.txt
cachetools==4.2.2
google-api-python-client==1.7.11
google-auth==1.6.3
google-auth-httplib2==0.0.3
httplib2==0.13.1
oauth2client==4.1.3
pyasn1==0.4.6
pyasn1-modules==0.2.6
rsa==4.0
six==1.12.0
uritemplate==3.0.0
django==3.2.9
python-dotenv==0.20.0
psycopg2-binary==2.8.6
gunicorn==20.0.4
python-telegram-bot==13.7
pycbrf==1.1.0
